fixed the memory replay buffer to adapt for continous values from the normalization

+++ roble/infrastructure/memory_optimized_replay_buffer.py
@@ -219,7 +219,7 @@ class MemoryOptimizedReplayBuffer(object):
         """
         if self.obs is None:
             self.obs = np.empty(
-                [self._size] + list(frame.shape), dtype=np.uint8
+                [self._size] + list(frame.shape), dtype=np.float16
             )  # dtype=np.float32 if self._lander else )
             self.obs_velocities = np.empty([self._size] + list(velocities.shape), dtype=np.float32)
             ## If discrete actions then just need a list of integers
